+++ index.ios.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native';
-import App from './App';
+import App from './js/App';
 
 AppRegistry.registerComponent('zestPlayground', () => App);
