+++ js/components/contentBrowser/HeaderTab.js
