+++ js/components/contentBrowser/HeaderTabChild.js
